move logic to selector
